moved ContactMe to separate component
